+++ lol_server_status/__init__.py
@@ -22,7 +22,7 @@ __prj__ = 'LoL Server Status'
 __author__ = 'Daniel Luque'
 __mail__ = 'danielluque14 at gmail dot com'
 __source__ = 'http://github.com/LuqueDaniel/LoL-Server-Status'
-__version__ = '1.1'
+__version__ = '1.2'
 __license__ = 'GPLv3'
 __docu__ = """LoL Server Status is an application for checking servers status of
 League of Legends"""
@@ -32,4 +32,4 @@ def run():
     #Import lol_server_status CORE
     from lol_server_status import core
 
-    core.run_lol_status_server()
+    core.run_lol_server_status()