+++ changelog.md
@@ -1,6 +1,13 @@
 LoL Server Status - Changelog
 ====================================================
 
+- **LoL Server Status 1.2** - (11/13/2013)
+	- LoL Server Status now opens a new tab in the browser with the server status forum.
+	- Now LoL Server Status can be compiled with cx_Freeze.
+	- Added new styles for QToolTip.
+	- Added new styles for QPushButton.
+	- Several improvements.
+
 - **LoL Server Status 1.1** - (07/21/2013)
 	- Added support for LAN server.
 	- Added support for LAS server.